requirements.txt
pandas>=1.5.0
google-cloud-bigquery>=3.4.0
openpyxl>=3.0.0
numpy>=1.21.0
requests>=2.28.0
msal>=1.20.0